Automatic "greedy" playing
Trying to get around full tree searches. When in doubt we just
end the move the "obvious" way (by playing all cards and buying
according to a fixed priority list). This gives it actual wiggle
room to make strategic moves, even though there is no real tactical
depth. Maybe we'll be able to train a network on in-move states
at some point down the line to fix this...

For the moment, the training set generation rate is still somewhat
low. Might need to optimise some more.