Optimise trade row ratings and redraw evaluation
The trade row forces us to evaluate a lot of very similar states
relative to each other - so introduce special code that only
does the conversion to a vector once, then makes the most out of
playing with the vector directly. This means we can send much
larger chunks to the neural network, and can skip a lot of
in-between logic.

Might somewhat decrease opportunities for caching though.