Support trace generation
The idea is that if we log how we got to an evaluation in the training
set from model data, we can cheaply repeat the run with a new model.
This obviously won't re-think all the tactical decisions that went into
the game-play, but this should still allow us to converge towards fixpoints
faster.