Use sigmoid on network's output, add layer
The sigmoid is pretty much exactly what value_to_prob & co were doing. Using the
network for this down-weighs extreme states, which has pros and cons. On the
positive side, this means we can dodge the question of whether won and lost
games do morally have an infinite rating. On the negative side, this likely
means that the network is going to start acting erratically for extreme game
states.