Small tree search improvements
Consider buying stuff even if it's not currently on the trade row (this was
actually only meant to be active temporarily).

Fill trade row with blob wheel(s) at end of finish_move_det. This ensures that
a neural network bias towards empty trade rows doesn't screw with us. The blob
wheel is relatively neutral and appears often enough in the deck to not mess
things up too badly.

Some logging changes.